chore: used LLONG_ instead of LONG_LONG_ in limits module. IntegerUnderlying is now defined as "long long" instead of "long"

+++ argon/vm/datatype/integer.h
@@ -11,7 +11,7 @@
 #include <argon/vm/datatype/decimal.h>
 
 namespace argon::vm::datatype {
-    using IntegerUnderlying = long;
+    using IntegerUnderlying = long long;
     using UIntegerUnderlying = unsigned long long;
 
     struct Integer {
